requirements.txt
FastAPI
sqlmodel
python-dotenv
